+++ dask/array/tests/test_array_core.py
@@ -3136,3 +3136,11 @@ def test_empty_chunks_in_array_len():
 
     err_msg = 'len() of unsized object'
     assert err_msg in str(exc_info.value)
+
+
+@pytest.mark.parametrize('dtype', [None, [('a', 'f4'), ('b', object)]])
+def test_meta(dtype):
+    a = da.zeros((1,), chunks=(1,))
+    assert a._meta.dtype == a.dtype
+    assert isinstance(a._meta, np.ndarray)
+    assert a.nbytes < 1000