+++ docs/source/changelog.rst
@@ -11,6 +11,7 @@ DataFrame
 +++++++++
 
 - Add ``dd.tseries.Resampler.agg`` (:pr:`3202`) `Richard Postelnik`_
+- Support operations that mix dataframes and arrays (:pr:`3230`) `Matthew Rocklin`_
 
 Bag
 +++